// src/api/dto/questionWithAnswerTranslationsItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Genshin Quiz API
 * API for Genshin Impact Quiz Application
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionTranslation } from './questionTranslation';

export type QuestionWithAnswerTranslationsItem = {
  language: string;
  text: QuestionTranslation;
};